Overrided volume keys to always increase the volume

+++ QNL/app/src/main/java/net/imanolgomez/qnl/BasicElement.java
@@ -18,7 +18,7 @@ public class BasicElement {
     // Instance variables
     private  int mId;
     private  String mName;
-    private  float mVersion;
+    private  double mVersion;
 
     /**
      * @param id The RouteElement's request ID
@@ -28,7 +28,7 @@ public class BasicElement {
     public BasicElement(
             int id,
             String name,
-            float version) {
+            double version) {
 
         // Set the instance fields from the constructor
         this.mId = id;
@@ -44,11 +44,11 @@ public class BasicElement {
         return mName;
     }
 
-    public float getVersion() {
+    public double getVersion() {
         return mVersion;
     }
 
-    public void setVersion(float version){
+    public void setVersion(double version){
         this.mVersion = version;
     }
