+++ QNL/app/src/main/java/net/imanolgomez/qnl/SoundManager.java
@@ -190,12 +190,12 @@ public class SoundManager {
         }
      }
 
-    public void playSample(int sampleId, float volume) {
+    public void playSample(int sampleId, double volume) {
         playSample(sampleId, volume, true);
     }
 
 
-    public void playSample(int sampleId, float volume, boolean setLooping) {
+    public void playSample(int sampleId, double volume, boolean setLooping) {
 
         stop();
 
@@ -237,7 +237,7 @@ public class SoundManager {
     public void addDefaultSample(){
 
         int id = -1; //The default sample is the non valid -1
-        float version = 1;
+        double version = 1;
         String name = DEFAULT_SAMPLE_NAME;
 
         BasicElement basicElement = new BasicElement(id,name,version);
